requirements.txt

google-cloud-storage
tqdm
PyWavelets
numpy
pandas
pytorch-lightning
pytest

# torch
# torchvision